+++ packages/client/src/components/leaderboard-list/leaderboard-list.tsx
@@ -1,28 +1,42 @@
 import Position from '../position';
 import './leaderboard-list.scss';
-import {LeaderboardSelectors} from "@/store/slices/leaderboard-slice";
-import {useSelector} from "react-redux";
+import { LeaderboardSelectors } from '@/store/slices/leaderboard-slice';
+import { useSelector } from 'react-redux';
+import { useEffect, useState } from 'react';
+import { getLeaderboardDatas } from '@/controllers/lider-controller';
+import { IQuery } from '@/api/types';
+import LeaderboardItem from './leaderboard-item';
 
 const LeaderboardList = () => {
+  const [data, setData] = useState<any>([]);
+  const leaderboardListData = useSelector(LeaderboardSelectors.all);
 
-    const leaderboardListData = useSelector(LeaderboardSelectors.all);
-    console.log('leaderboardListData = ', leaderboardListData)
-    const copyArray = [...leaderboardListData.leaderboard];
-    const sortedArray = copyArray.sort((a, b) => (Number(b.score) - Number(a.score)));
+  useEffect(() => {
+    const query: IQuery = {
+      // сортировка score, 1 страница на 10 записей
+      ratingFieldName: 'score',
+      cursor: 0,
+      limit: 10,
+    };
 
-    const itemList = sortedArray.map((item, index) => {
-        const position = index + 1;
+    (async () => {
+      const res = await getLeaderboardDatas(query);
 
-        return (
-            <li className="leaderboard__row" key={position}>
-                <Position className="leaderboard__position" position={position}/>
-                <div className="leaderboard__name">{item.name}</div>
-                <div className="leaderboard__score">{item.score}</div>
-            </li>
-        );
-    });
+      setData(res);
+    })();
+  }, []);
 
-    return <ul className="leaderboard__data">{itemList}</ul>;
+  console.log('leaderboardListData = ', leaderboardListData);
+  const copyArray = [...leaderboardListData.leaderboard];
+  const sortedArray = copyArray.sort((a, b) => Number(b.score) - Number(a.score));
+
+  return (
+    <ul className="leaderboard__data">
+      {sortedArray.map((item, index) => (
+        <LeaderboardItem position={index + 1} item={item} />
+      ))}
+    </ul>
+  );
 };
 
 export default LeaderboardList;
